fix: bitfield width, year encoding, defer flash write out of ISR
Bug 1: movement_active_hours_t had 'uint8_t reserved : 17' which is
invalid — uint8_t can only hold 8 bits. Changed to uint32_t to match
the 17-bit reservation and the 32-bit backing register.

Bug 2: encode_date() compared dt.unit.year (offset-from-2020, 0-63)
against the absolute year 2024 — always false for any valid date.
Fixed to compare against offset 4 (2020+4=2024). Also fixes the
'year always zero in date_code' issue reported in PR #5.

Bug 3: active_hours advance callbacks called movement_set_active_hours()
(-> watch_store_backup_data()) directly on every button press. Added
module-level deferred-write pattern: advance functions modify a local
pending copy and set s_active_hours_dirty; the actual write is flushed
on the next EVENT_TICK and when the face resigns.

+++ movement.h
@@ -109,7 +109,7 @@ typedef union {
         uint8_t start_quarter_hours : 7;  // 0-95 (0=00:00, 95=23:45, 15-min increments)
         uint8_t end_quarter_hours : 7;    // 0-95 (0=00:00, 95=23:45, 15-min increments)
         bool enabled : 1;                  // active hours enabled/disabled
-        uint8_t reserved : 17;             // reserved for future use
+        uint32_t reserved : 17;            // reserved for future use
     } bit;
     uint32_t reg;
 } movement_active_hours_t;
